Refactor: Move path grammar to paths.js
