+++ packages/utils/src/objects.js
@@ -1,7 +1,7 @@
 import { clone } from './cloning.js';
 import { isEqual } from './equality.js';
-import { configured } from './functions.js';
 import { each } from './loops.js';
+import { eachPath, elementKey, get, keyValueFrom } from './paths.js';
 import { escapeRegExp } from './regexp.js';
 import { isArray, isMap, isObject, isPlainObject, isSet, isString } from './types.js';
 
@@ -80,59 +80,6 @@ const overBudget = (value) => spendBudget(value, autoBudget) < 0;
   equality and report their own path. A wholesale change to a non-container
   root reports path '' (the RFC 6902 root convention).
 */
-// the identity fields a keyed array element is matched on, first present wins —
-// identity of an array element, shared by every keyed helper, vocabulary adjustable at boot
-export const elementKey = /* @__PURE__ */ configured(
-  (item, keys = elementKey.config.keys) => {
-    if (!isObject(item)) {
-      return undefined;
-    }
-    // fast track the default vocabulary, a static chain instead of the field loop
-    if (keys.length === 4 && keys[0] === 'id' && keys[1] === '_id' && keys[2] === 'hash' && keys[3] === 'key') {
-      return item.id ?? item._id ?? item.hash ?? item.key;
-    }
-    for (const field of keys) {
-      if (item[field] != null) {
-        return item[field];
-      }
-    }
-    return undefined;
-  },
-  { keys: ['id', '_id', 'hash', 'key'] },
-);
-
-// resolve a keyed array segment to a live index, -1 if absent. identity compares
-// String-coerced because the path carries a string and ids may be number or string
-const findKeyed = (array, keyValue, keys = elementKey.config.keys) => {
-  if (!isArray(array)) {
-    return -1;
-  }
-  for (let i = 0; i < array.length; i++) {
-    const id = elementKey(array[i], keys);
-    if (id != null && String(id) === keyValue) {
-      return i;
-    }
-  }
-  return -1;
-};
-
-// a key value can ride the dot-bracket path grammar unless it carries ']', the
-// one character that closes a keyed selector. dots, '@', '[' and a leading '#'
-// all round-trip (items[#jack@semantic-ui.com]). shared by the keyed diff and
-// the keyed read paths so an emitted path always parses back through get
-const isKeyedPathSafe = (keyValue) => keyValue.indexOf(']') === -1;
-
-// an element's identity spelled for a path, null when the element is unkeyed
-// or its key can't ride the grammar
-const keyValueFrom = (element, keys) => {
-  const id = elementKey(element, keys);
-  if (id == null) {
-    return null;
-  }
-  const keyValue = String(id);
-  return isKeyedPathSafe(keyValue) ? keyValue : null;
-};
-
 // index a keyed array by stringified identity for set-wise diffing. returns null
 // (caller falls back to the positional walk) when an element is unkeyed, two share
 // a key, or a key value can't ride the dot-bracket path grammar — so detectChanges
@@ -267,40 +214,6 @@ const pruneChildPaths = (pathLog) => {
     }
   }
   return pruned;
-};
-
-/*
-  Walk the paths a path passes through, shortest first — 'todos[#a].done'
-  visits 'todos', 'todos[#a]', 'todos[#a].done'. Every visited value is itself
-  a resolvable path: a bracket splits its segment into container and element,
-  and a dot inside a keyed id is identity, never a boundary. Pass self: false
-  to visit only the ancestors above the target. Returning false stops the
-  walk, like each.
-*/
-export const eachPath = (path, callback, { self = true } = {}) => {
-  let index = 0;
-  for (let i = 0; i < path.length; i++) {
-    const char = path[i];
-    if (char === '[') {
-      if (i > 0 && callback(path.slice(0, i), index++, path) === false) {
-        return path;
-      }
-      const close = path.indexOf(']', i + 1);
-      if (close === -1) {
-        break;
-      }
-      i = close;
-    }
-    else if (char === '.') {
-      if (i > 0 && callback(path.slice(0, i), index++, path) === false) {
-        return path;
-      }
-    }
-  }
-  if (self && path !== '') {
-    callback(path, index, path);
-  }
-  return path;
 };
 
 // the read mirror of pruneChildPaths: a deeper read subsumes its ancestors.
@@ -924,443 +837,6 @@ export const arrayFromObject = (obj) => {
     });
   });
   return arr;
-};
-
-const INDEX_SEGMENT = /^\d+$/;
-const isIndexSegment = (part) => INDEX_SEGMENT.test(part);
-
-// a bracket segment resolves to a keyed selector (field[#identity]) or a
-// numeric positional index (field[2]). null for a malformed segment — an
-// unclosed bracket, or a positional body that isn't a whole index — which
-// every caller treats as addressing nothing
-const extractBracketAccess = (part) => {
-  const bracketIndex = part.indexOf('[');
-  const closeIndex = part.indexOf(']', bracketIndex + 1);
-  if (closeIndex === -1) {
-    return null;
-  }
-  const key = part.substring(0, bracketIndex);
-  const body = part.substring(bracketIndex + 1, closeIndex);
-  // a leading # marks a keyed element selector; the identity rides as a raw string value
-  if (body.charCodeAt(0) === 35) {
-    return { key, keyValue: body.slice(1) };
-  }
-  if (!isIndexSegment(body)) {
-    return null;
-  }
-  return { key, index: parseInt(body, 10) };
-};
-
-// the segment split under every path walk. a dot inside a bracket belongs to
-// the identity (items[#jack@semantic-ui.com].role), so '[' fast-forwards to
-// its ']' and only outside dots separate segments
-const splitPath = (path) => {
-  if (path.indexOf('[') === -1) {
-    return path.split('.');
-  }
-  const parts = [];
-  let start = 0;
-  for (let i = 0; i < path.length; i++) {
-    const char = path[i];
-    if (char === '[') {
-      const close = path.indexOf(']', i + 1);
-      if (close === -1) {
-        break;
-      }
-      i = close;
-    }
-    else if (char === '.') {
-      parts.push(path.slice(start, i));
-      start = i + 1;
-    }
-  }
-  parts.push(path.slice(start));
-  return parts;
-};
-
-// the one walk under get and has. resolves the dot-bracket grammar to the
-// stored value, MISSING when the path never lands on one. a stored undefined
-// resolves as undefined, so has can tell it apart from a missing path
-const MISSING = Symbol('missing');
-
-const resolvePath = (obj, path, keys) => {
-  if (obj === null || !isObject(obj)) {
-    return MISSING;
-  }
-
-  const parts = splitPath(path);
-  let currentObject = obj;
-  let pathOffset = 0;
-
-  for (let i = 0; i < parts.length; i++) {
-    if (currentObject === null || !isObject(currentObject)) {
-      return MISSING;
-    }
-
-    const part = parts[i];
-
-    if (part.includes('[')) {
-      const access = extractBracketAccess(part);
-      if (access === null) {
-        return MISSING;
-      }
-      // an empty key ([0].x against an array root) addresses the current value as the array
-      const array = access.key === '' ? currentObject : currentObject[access.key];
-      if (!isArray(array)) {
-        return MISSING;
-      }
-      const index = 'keyValue' in access ? findKeyed(array, access.keyValue, keys) : access.index;
-      if (index < 0 || index >= array.length) {
-        return MISSING;
-      }
-      currentObject = array[index];
-    }
-    else if (part in currentObject) {
-      currentObject = currentObject[part];
-    }
-    else {
-      // an existing literal dotted key wins, the whole remainder first (obj['a.b.c'])
-      const remainingPath = path.substring(pathOffset);
-      if (remainingPath in currentObject) {
-        currentObject = currentObject[remainingPath];
-        break;
-      }
-      // then this and the next part combined (obj['a.b'])
-      const combinedKey = `${part}.${parts[i + 1]}`;
-      if (combinedKey in currentObject) {
-        currentObject = currentObject[combinedKey];
-        i++;
-        // the loop footer advances past the first part only, cover the consumed second
-        pathOffset += parts[i].length + 1;
-      }
-      else {
-        return MISSING;
-      }
-    }
-
-    pathOffset += part.length + 1;
-  }
-
-  return currentObject;
-};
-
-/*
-  Access a nested object field from a string, like 'a.b.c'
-*/
-export const get = function(obj, path = '', keys = elementKey.config.keys) {
-  if (typeof path !== 'string') {
-    return undefined;
-  }
-  // simple property access — no dots, no brackets
-  if (path.indexOf('.') === -1 && path.indexOf('[') === -1) {
-    return (obj !== null && isObject(obj)) ? obj[path] : undefined;
-  }
-  const value = resolvePath(obj, path, keys);
-  return value === MISSING ? undefined : value;
-};
-
-/*
-  Existence twin of get — true when the path resolves to a real location,
-  even one holding undefined. get() conflates a missing path with a stored
-  undefined, has() is how callers tell them apart.
-*/
-export const has = function(obj, path = '', keys = elementKey.config.keys) {
-  if (typeof path !== 'string') {
-    return false;
-  }
-  // simple property access — no dots, no brackets
-  if (path.indexOf('.') === -1 && path.indexOf('[') === -1) {
-    return obj !== null && isObject(obj) && path in obj;
-  }
-  return resolvePath(obj, path, keys) !== MISSING;
-};
-
-// set and unset act only on a real path into a real object, and refuse
-// prototype pollution (__proto__ and friends)
-const CLIMBS_PROTOTYPE = /(^|\.|\[)(__proto__|constructor|prototype)(\.|\[|\]|$)/;
-
-// a keyed id is an opaque identity, never a property access, so guards that
-// read path structure see it blanked (users[#mail.constructor.dev].role passes)
-const KEYED_ID = /\[#[^\]]*\]/g;
-const pathShape = (path) => path.indexOf('[#') === -1 ? path : path.replace(KEYED_ID, '[]');
-
-const isWritablePath = (obj, path) =>
-  typeof path === 'string' && path !== '' && obj !== null && isObject(obj)
-  && !CLIMBS_PROTOTYPE.test(pathShape(path));
-
-/*
-  Set a nested object field from a string path, the write twin of get, so
-  paths from trackWrites and detectChanges apply back. Creates missing
-  intermediates — arrays when the next segment is an index, objects otherwise.
-*/
-export const set = function(obj, path, value, keys = elementKey.config.keys) {
-  if (!isWritablePath(obj, path)) {
-    return obj;
-  }
-
-  // simple property access — no dots, no brackets
-  if (path.indexOf('.') === -1 && path.indexOf('[') === -1) {
-    obj[path] = value;
-    return obj;
-  }
-
-  const parts = splitPath(path);
-  let currentObject = obj;
-  let pathOffset = 0;
-
-  for (let i = 0; i < parts.length; i++) {
-    const part = parts[i];
-    const isLast = i === parts.length - 1;
-
-    if (part.includes('[')) {
-      const access = extractBracketAccess(part);
-      if (access === null) {
-        return obj;
-      }
-      let array;
-      if (access.key === '') {
-        // an empty key addresses the current value as the array; it can't be vivified in place
-        if (!isArray(currentObject)) {
-          return obj;
-        }
-        array = currentObject;
-      }
-      else {
-        if (!isArray(currentObject[access.key])) {
-          currentObject[access.key] = [];
-        }
-        array = currentObject[access.key];
-      }
-      if ('keyValue' in access) {
-        const index = findKeyed(array, access.keyValue, keys);
-        if (isLast) {
-          if (index === -1) {
-            array.push(value); // keyed add: an absent key appends a new element
-          }
-          else {
-            array[index] = value; // keyed replace: a present key overwrites in place
-          }
-          return obj;
-        }
-        if (index === -1) {
-          return obj; // a field write to a vanished element is a no-op
-        }
-        currentObject = array[index];
-      }
-      else {
-        const { index } = access;
-        if (isLast) {
-          array[index] = value;
-          return obj;
-        }
-        if (array[index] === null || !isObject(array[index])) {
-          array[index] = isIndexSegment(parts[i + 1]) ? [] : {};
-        }
-        currentObject = array[index];
-      }
-    }
-    else if (isLast) {
-      currentObject[part] = value;
-      return obj;
-    }
-    else {
-      if (!(part in currentObject)) {
-        // an existing literal dotted key wins over creating intermediates,
-        // mirroring get()'s resolution
-        const remainingPath = path.substring(pathOffset);
-        if (remainingPath in currentObject) {
-          currentObject[remainingPath] = value;
-          return obj;
-        }
-        const combinedKey = `${part}.${parts[i + 1]}`;
-        if (combinedKey in currentObject) {
-          if (currentObject[combinedKey] === null || !isObject(currentObject[combinedKey])) {
-            currentObject[combinedKey] = isIndexSegment(parts[i + 2]) ? [] : {};
-          }
-          currentObject = currentObject[combinedKey];
-          pathOffset += combinedKey.length + 1;
-          i++;
-          continue;
-        }
-      }
-      if (currentObject[part] === null || !isObject(currentObject[part])) {
-        currentObject[part] = isIndexSegment(parts[i + 1]) ? [] : {};
-      }
-      currentObject = currentObject[part];
-    }
-
-    pathOffset += part.length + 1;
-  }
-
-  return obj;
-};
-
-/*
-  Remove a nested object field from a string path, the delete twin of get/set.
-  A missing path is a no-op, and a removed array index leaves a hole rather
-  than splicing, so sibling index paths stay valid when applying several
-  removals at once.
-*/
-export const unset = function(obj, path, keys = elementKey.config.keys) {
-  if (!isWritablePath(obj, path)) {
-    return obj;
-  }
-
-  // simple property access — no dots, no brackets
-  if (path.indexOf('.') === -1 && path.indexOf('[') === -1) {
-    delete obj[path];
-    return obj;
-  }
-
-  const parts = splitPath(path);
-  let currentObject = obj;
-  let pathOffset = 0;
-
-  for (let i = 0; i < parts.length; i++) {
-    const part = parts[i];
-    const isLast = i === parts.length - 1;
-
-    if (part.includes('[')) {
-      const access = extractBracketAccess(part);
-      if (access === null) {
-        return obj;
-      }
-      const array = access.key === '' ? currentObject : currentObject[access.key];
-      if (!isArray(array)) {
-        return obj;
-      }
-      if ('keyValue' in access) {
-        const index = findKeyed(array, access.keyValue, keys);
-        if (index === -1) {
-          return obj;
-        }
-        if (isLast) {
-          array.splice(index, 1); // keyed removal splices: no positional siblings to keep valid
-          return obj;
-        }
-        currentObject = array[index];
-      }
-      else {
-        const { index } = access;
-        if (isLast) {
-          delete array[index];
-          return obj;
-        }
-        currentObject = array[index];
-      }
-    }
-    else if (isLast) {
-      delete currentObject[part];
-      return obj;
-    }
-    else if (part in currentObject) {
-      currentObject = currentObject[part];
-    }
-    else {
-      // an existing literal dotted key wins, mirroring get()'s resolution
-      const remainingPath = path.substring(pathOffset);
-      if (remainingPath in currentObject) {
-        delete currentObject[remainingPath];
-        return obj;
-      }
-      const combinedKey = `${part}.${parts[i + 1]}`;
-      if (combinedKey in currentObject) {
-        currentObject = currentObject[combinedKey];
-        if (currentObject === null || !isObject(currentObject)) {
-          return obj;
-        }
-        pathOffset += combinedKey.length + 1;
-        i++;
-        continue;
-      }
-      return obj;
-    }
-
-    if (currentObject === null || !isObject(currentObject)) {
-      return obj;
-    }
-    pathOffset += part.length + 1;
-  }
-
-  return obj;
-};
-
-// a path can address an identity-bearing array element two ways: by position (items.0.qty) or by
-// key (items[#a].qty). the positional spelling is only stable while the array doesn't move;
-// keyedPath resolves it to the keyed spelling against a live object, rewriting each positional
-// segment whose element carries a path-safe identity. returns the input string itself when
-// nothing rewrites, so callers compare by reference
-const HAS_POSITIONAL_SEGMENT = /(^|\.|\[)\d/;
-
-export const keyedPath = (obj, path, keys = elementKey.config.keys) => {
-  // the probe reads pathShape so a digit inside a keyed id (items[#200.40.50])
-  // doesn't look like a positional segment
-  if (typeof path !== 'string' || !HAS_POSITIONAL_SEGMENT.test(pathShape(path)) || obj === null || !isObject(obj)) {
-    return path;
-  }
-  const parts = splitPath(path);
-  // a leading index means the root itself is an array; get/set cannot parse a leading bracket, so
-  // that spelling stays positional
-  if (isIndexSegment(parts[0])) {
-    return path;
-  }
-  const out = [];
-  let currentObject = obj;
-  let changed = false;
-  for (let i = 0; i < parts.length; i++) {
-    const part = parts[i];
-    if (currentObject === null || !isObject(currentObject)) {
-      return path;
-    }
-    if (part.includes('[')) {
-      const access = extractBracketAccess(part);
-      if (access === null) {
-        return path;
-      }
-      const array = currentObject[access.key];
-      if (!isArray(array)) {
-        return path;
-      }
-      const index = 'keyValue' in access ? findKeyed(array, access.keyValue, keys) : access.index;
-      const element = array[index];
-      if ('keyValue' in access) {
-        out.push(part);
-      }
-      else {
-        const keyValue = keyValueFrom(element, keys);
-        if (keyValue !== null) {
-          out.push(`${access.key}[#${keyValue}]`);
-          changed = true;
-        }
-        else {
-          out.push(part);
-        }
-      }
-      currentObject = element;
-    }
-    else if (isIndexSegment(part) && isArray(currentObject)) {
-      const element = currentObject[Number(part)];
-      const keyValue = keyValueFrom(element, keys);
-      if (keyValue !== null) {
-        out[out.length - 1] += `[#${keyValue}]`;
-        changed = true;
-      }
-      else {
-        out.push(part);
-      }
-      currentObject = element;
-    }
-    else if (part in currentObject) {
-      out.push(part);
-      currentObject = currentObject[part];
-    }
-    else {
-      // an unresolvable segment (including get()'s literal dotted-key forms) holds no positional
-      // rewrite; leave the spelling alone
-      return path;
-    }
-  }
-  return changed ? out.join('.') : path;
 };
 
 /* This is useful for callbacks or other scenarios where you want to avoid the
